Further work on routing
