+++ java/code/src/com/suse/manager/webui/templates/schedule/maintenance-window-details.jade
@@ -1,0 +1,16 @@
+include /templates/common.jade
+
+#maintenance-window-details
+
+script(type='text/javascript').
+    window.csrfToken = "#{csrf_token}";
+    window.timezone = "#{h.renderTimezone()}";
+    window.localTime = "#{h.renderLocalTime()}";
+
+script(type='text/javascript').
+    spaImportReactPage('maintenance/maintenance-window-details')
+        .then(function (module) { module.renderer('maintenance-window-details', {
+            type: "#{type}",
+            isAdmin: #{isAdmin},
+        })
+    });
