+++ java/code/src/com/suse/manager/webui/templates/schedule/maintenance-windows-list.jade
@@ -1,0 +1,16 @@
+include /templates/common.jade
+
+#maintenance-windows-list
+
+script(type='text/javascript').
+    window.csrfToken = "#{csrf_token}";
+    window.timezone = "#{h.renderTimezone()}";
+    window.localTime = "#{h.renderLocalTime()}";
+
+script(type='text/javascript').
+    spaImportReactPage('maintenance/maintenance-windows-list')
+        .then(function (module) { module.renderer('maintenance-windows-list', {
+            type: "#{type}",
+            isAdmin: #{isAdmin},
+        })
+    });
