--- java/code/src/com/suse/manager/webui/templates/schedule/maintenance-windows.jade
@@ -1,14 +1,0 @@
-include /templates/common.jade
-
-#maintenance-windows
-
-script(type='text/javascript').
-    window.csrfToken = "#{csrf_token}";
-    window.timezone = "#{h.renderTimezone()}";
-    window.localTime = "#{h.renderLocalTime()}";
-    window.type = "#{type}";
-    window.isAdmin = "#{isAdmin}";
-
-script(type='text/javascript').
-    spaImportReactPage('maintenance/maintenance-windows')
-        .then(function (module) { module.renderer('maintenance-windows') });
