+++ web/html/src/manager/maintenance/index.ts
@@ -1,4 +1,5 @@
 export default {
-  "maintenance/maintenance-windows": () => import("./maintenance-windows"),
+  "maintenance/maintenance-windows-list": () => import("./list/maintenance-windows-list"),
+  "maintenance/maintenance-window-create-edit": () => import("./edit/maintenance-window-create-edit"),
   "maintenance/system-assignment": () => import("./ssm/system-assignment"),
 };
